requirements.txt
matplotlib
tensorflow
streamlit
pillow
numpy